requirements.txt
colorama
jsonschema
optparse-pretty
pillow
PyYAML
requests
samsungtvws[async,encrypted]
tomlkit
urllib3
wakeonlan
